add test unitaire

+++ aymeric/evenement.py
@@ -9,6 +9,11 @@ class Evenements:
         self.username = username
 
     def event_already_exists(self, date, name):
+        """Si un évènement existe
+
+        PRE : Reçoit une date, un nom, un username 
+        POST : Si les arguments reçus correspondent à un évènement qui existe déjà renvoi True
+        """
         ret = False
         if os.path.exists(self.FILENAME_EVENTS):
             with open(self.FILENAME_EVENTS, 'r') as csvfile:
@@ -20,6 +25,11 @@ class Evenements:
         return ret
 
     def add_event(self, date, name, acteurs):
+        """Ajout d'évènement
+
+        PRE : Reçoit une date, un nom, un username 
+        POST : Si les arguments reçus ne correspondent à aucun évènement qui existe déjà, renvoi True et le créé l'évènement
+        """
         ret = False
         if os.path.exists(self.FILENAME_EVENTS):
             with open(self.FILENAME_EVENTS, "a", newline='') as csvfile:
@@ -34,6 +44,11 @@ class Evenements:
         return ret
 
     def get_event(self, date, name):
+        """Récuperer un évènement
+
+        PRE : Reçoit une date, un nom, un username 
+        POST : Si les arguments reçus correspondent à un évènement qui existe déjà renvoi True
+        """
         e = None
         if os.path.exists(self.FILENAME_EVENTS):
             with open(self.FILENAME_EVENTS, 'r') as csvfile:
@@ -68,17 +83,18 @@ class Evenements:
     def remove_event(self, date, name):
         lines = []
         n = 0
-        with open(self.FILENAME_EVENTS, "r") as csvfile:
-            reader = csv.reader(csvfile)
-            for row in reader:
-                if len(row):
-                    if row[0] != date or row[1] != name or row[2] != self.username:
-                        lines.append(row)
-                    else:
-                        n += 1
-        with open(self.FILENAME_EVENTS, "w", newline='') as csvfile:
-            writer = csv.writer(csvfile)
-            writer.writerows(lines)
+        if os.path.exists(self.FILENAME_EVENTS):
+            with open(self.FILENAME_EVENTS, "r") as csvfile:
+                reader = csv.reader(csvfile)
+                for row in reader:
+                    if len(row):
+                        if row[0] != date or row[1] != name or row[2] != self.username:
+                            lines.append(row)
+                        else:
+                            n += 1
+            with open(self.FILENAME_EVENTS, "w", newline='') as csvfile:
+                writer = csv.writer(csvfile)
+                writer.writerows(lines)
         return n > 0
